requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-active-link==0.2.2
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-js-asset==3.1.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
